Remove the updates generation parameters
This commit changes the way update requests are generated.  They used to
be generated up to a certain number.  Now as many updates as the user
wants can be generated, using infinite streams.

This not only simplifies the configuration of update-related benchmarks
but also reduce the amount of boilerplate code.

Typically the `BufferedRandomIterator` class was needed to shuffle
updates, as otherwise, multiple updates against the same entity would be
sent simultaneously.  Now, before an entity is updated again, all the
other entities of the dataset need to receive an update as well.  This
removes the need for the shuffling and buffering operation.

As a result, there are no random operation that are performed anymore in
the simulations.  The seed parameter is not necessary anymore.  It has
been removed as it is dead code.